# PERF: Switch conditions in RangeIndex._shallow_copy