xref https://github.com/pandas-dev/pandas/pull/57534#issuecomment-1957832841

`is_range_indexer` lets us short circuit unlike checking the remainder

```python
In [1]: from pandas import *; import numpy as np
   ...: np.random.seed(123)
   ...: size = 1_000_000
   ...: ngroups = 1000
   ...: data = Series(np.random.randint(0, ngroups, size=size))
+ /opt/miniconda3/envs/pandas-dev/bin/ninja
[1/1] Generating write_version_file with a custom command

In [2]: %timeit data.groupby(data).groups
16.8 ms ± 183 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  # PR

In [3]: %timeit data.groupby(data).groups
17.6 ms ± 114 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  # main
```